Banner Management updates

+++ admin/delete_banner.php
@@ -3,17 +3,33 @@
 require 'config.php';
 include '../module/Application/src/Model/include_dao.php';
 
+$response = [
+	'status' => false,
+	'msg' => 'Error',
+];
+
 $bannerMySqlExtDAO = new BannerMySqlExtDAO();
+$bannerImageMySqlExtDAO = new BannerImageMySqlExtDAO();
 
 if (isset($_REQUEST['id'])) {
 	$id = $_REQUEST['id'];
-	$banner = $bannerMySqlExtDAO->load($id);
-	print_r(is_file(IMAGES_PATH . $banner->image));
-	if (is_file(IMAGES_PATH . $banner->image)) {
-		unlink(IMAGES_PATH . $banner->image);
-		unlink(IMAGES_PATH . "med_" . $banner->image);
-		unlink(IMAGES_PATH . "small_" . $banner->image);
+	$hasChildren = $bannerImageMySqlExtDAO->queryByBannerId($id);
+	if ($hasChildren) {
+		$response['msg'] = "This banner has images, delete them first";
+	} else {
+		$banner = $bannerMySqlExtDAO->load($id);
+		if (is_file(IMAGES_PATH . $banner->image)) {
+			unlink(IMAGES_PATH . $banner->image);
+			unlink(IMAGES_PATH . "med_" . $banner->image);
+			unlink(IMAGES_PATH . "small_" . $banner->image);
+		}
+		$delete = $bannerMySqlExtDAO->delete($id);
+		if ($delete) {
+			$response = [
+				'status' => true,
+				'msg' => 'Deleted',
+			];
+		}
 	}
-	$delete = $bannerMySqlExtDAO->delete($id);
-	exit($delete);
 }
+echo json_encode($response);
